requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
# Optional: only needed if USE_PLAYWRIGHT_FOR_TOKEN=true
# playwright>=1.40.0
